requirements.txt
Django==1.3.1
django-tinymce==1.5.1a2
requests==1.2.0
django-uuidfield==0.4.0
